+++ src/locale/de-de.js
@@ -10,12 +10,10 @@ const numbers = {
 	romanOrdinalIndicator :		"\\.",
 }
 
-const abbreviationsForNbsp = [
-	"S.",
-	"z."
+const singleWordAbbreviations = [
+	"S",
+	"z"
 ]
-
-const singleWordAbbreviations = []
 
 // Disregard correct spelling in your locale
 // and provide abbreviations in format “abbr abbr abbr…”
@@ -50,7 +48,6 @@ const multipleWordAbbreviations = [
 export default {
 	quotes,
 	numbers,
-	abbreviationsForNbsp,
 	singleWordAbbreviations,
 	multipleWordAbbreviations
 }