+++ src/locale/en-us.js
@@ -10,12 +10,12 @@ const numbers = {
 	romanOrdinalIndicator :		"",
 }
 
-const abbreviationsForNbsp = [
-	"p.",
-	"pp."
+const singleWordAbbreviations = [
+	"p",
+	"pp",
+	"no",
+	"vol"
 ]
-
-const singleWordAbbreviations = []
 
 // Disregard correct spelling in your locale
 // and provide abbreviations in format “abbr abbr abbr…”
@@ -30,7 +30,6 @@ const multipleWordAbbreviations = [
 export default {
 	quotes,
 	numbers,
-	abbreviationsForNbsp,
 	singleWordAbbreviations,
 	multipleWordAbbreviations
 }